tests: Fix DateUtil tests
Some of them were wrong by an hour because of incorrect timezone handling.
I've checked the correct timestamps with https://www.epochconverter.com/
and replaced the old values in the tests. Now that DateUtil uses `timegm()`
and doesn't use `timezone` we're guaranteed to always use GMT times.

Also, a new test has been added to verify that time is formatted correctly
for timestamps that aren't the timestamp 0.