scrollpane: Fix clash of palettes of toolbar and scroll sprite
Commit 6320eb6df688 ("scrollpane: Actually load palette of scroll
sprite") made sure to load the palette of the scroll sprite, but it
didn't notice that the sprite palette was already in use by the toolbar
image.

The toolbar image is a 256 color sprite that doesn't use all colors. We
can use the last 16 color palette for the scroll sprite safely,