image: Handle transparency in images correctly
Instead of setting the color to white while loading the image, skip drawing
pixels that have an alpha bit of 0. That way we don't need to worry about

Using PLUM_ALPHA_INVERT makes it so that bit 15 of the color is 1 if the
pixel is opaque, and 0 if it's transparent. This is consisten with the way
it works in the hardware. Also, this commit fixes transparency in images
that didn't have to be scaled down. They were copied with `memcpy()`
without checking the alpha bit.