html: tests: Fix code that feeds data to the HTML parser
The HTTP client feeds data to the Document as data is received. The
Document feeds it to the HeaderParser, which eventually feeds it to the
HtmlParser.

The problem is that nothing in this chain stores any data. The HTML parser
uses method "next()" to get the following character from the input (in
UTF-8 encoding, for example) and it isn't prepared to handle the case in
which next() doesn't return a valid character.

The HTTP client feeds data in chunks of arbitrary sizes, so the HTML parser
may receive data that leaves it in a loop without an exit. That loop will
keep calling "next()", which knows there is no more data available, and
the loop never ends.

This commit modifies the code so that all data fed from the cache or HTTP
clients is stored in a temporary buffer. When the full data has been read
from the cache or downloaded from the server it's safer to parse it.

This isn't a perfect solution. If the HTML received is malformed it's still
possible to get stuck in infinite loops. However, this patch fixes loops
caused by valid HTML.