canvas: Refactor constructor and destructor
m_bgMain and m_bgSub are now always created, even in the SDL port. This
will help us test that code in the SDL port, not just use it in NDS.

This also fixes potential memory leaks. The objects allocated in the
constructor weren't freed in the destructor (and the constructor didn't
check if they had been allocated correctly).