Component: Support statically-allocated children
Ideally we want everything to be deleted by the same class that created
it so that the ownership model is easier to understand. However, this
will take a while, so it's better to have a mixed model that supports
staticallu-allocated and dynamically-allocated instances..

Also, websites won't ever have a system in which all objects are
statically allocated, so we will need a mixed system for them.