tests: Fix cookie tests
The sdroot folder wasn't deleted after every test case, so some tests
failed because of the files generated by the previous tests.

Also, the CookieTest.Expires test failed because the timestamp used as
reference was incorrect. Some additional checks have been added to ensure
that the timestamp is the exact one that we need to test.