URI: tests: Fix file:// protocol
Previously we assumed that all paths such as "file:///file.txt" started
at the root of the SD card in the NDS port, or at the current working
directory in the SDL port.

It didn't support a server name. Paths like "file://localhost/file.txt"
weren't supported.

The problem with this is that the path to the application files of
Bunjalloo depends on the system. This is specified in the DATADIR
define, but it isn't very helpful in some cases.

This patch adds support for server names to the "file://" protocol, and
it adds support for "localhost" and "bunjalloo". If no server is
specified, "localhost" is used by default.

The root of "localhost" is the root of the filesystem in NDS (or the
current working dir in the SDL port). The root of "bunjalloo" is the
path "/data/bunjalloo/", wherever that is in your system.

All code that used file paths has been updated to support the two
servers, and the tests have been improved to test both having a server
name or not.