FAT: Remove sdroot folder from the repository
This folder is now created automatically, so it doesn't need to be present
in the repository. In fact, it's worse to keep it because anyone using the
SDL version will be creating files inside of it and they will need to be
ignored by developers that want to create commits.