ComboBox: Fix rendering of very big combo boxes
If options are longer than the screen width the scroll bar of the combo box
is currently drawn outside of the screen. This check limits the size of the
combo box to the width of the screen minus enough space to fit the scroll
bar of the combo box and the scroll bar of the website itself.