Font: ComboBox: Fix underflow in X bounds calculation
This can sometimes become a negative number, causing out-of-bound writes.

This could be triggered by a ComboBox with items that are too long to fit
on the screen. When that item was selected, it would crash (it wouldn't
crash if that item wasn't selected).

Now it just prints the text outside of the ComboBox bounds, but that's
better than a crash.