TextArea: TextContainer: Fix shortening of long words
When calling shorterWordFromLong() sometimes we would get a word that would
be one pixel too long to fit and it would end up being split between two
lines.

In the case of TextArea we need a bit extra because if not it looks weird.
There's always a bit of extra space on the left, so we need extra space on
the right to have a nice margin.