EditableTextArea: Fix addition and deletion of characters
Previously the code was very buggy, particularly when adding or removing
characters that shifted words from one line to a different one.

The new code has been simplified. Adding or deleting code reshuffles the
text whenever there are changes so that the final result is correct. It
tries to remove lots of checks from the previous code by simplifying it,
even if it becomes a bit less efficient. After all, when the user is
typing, that's the only thing that is going on, so we don't need the code
to be super fast, just correct.