EditableTextArea: Always place caret left of newline characters
This is only important for multiline input forms.

Leaving the caret on the right means that there is a weird space between
the last character and the caret (the space that the newline takes in the
font). This is counter-intuitive, and it also makes it so that the next
character is added to the following line.

This check moves the caret one character to the left if the selecter
character is a newline character.