DejaVu: FAT: Add script to convert the font and update it
The old font in NitroFS doesn't match the output of convertftf when the
Bitstream Vera font is provided as input. Also, Bitstream Vera is less
complete than DejaVu, so it's a good idea to replace the old font by a
new one that we can regenerate at any point.

This is also useful if we want to add support for bold and italics in
the future. It wasn't possible to obtain bold and italics fonts for Vera
because it wasn't possible to match the graphics.

The new font is a bit bigger, but not by much.