build: Refactor build system
Previously any modification to the code of libbunjalloo, libbwt or
libndspp would require a clean build of bunjalloo to get the code of the
new libraries. This is pretty inconvenient because it's very easy to
forget to rebuild the main binary and test the old build.

This patch moves all the code of bunjalloo and its libraries to the
"bunjalloo" folder. It also modifies the Makefile inside that folder to
create a static library in addition to the regular binary for the
platform.

It also deletes the Doxygen configuration files as they aren't used for
anything and they are outdated anyway.