tests: Fix updater test
This was missing a call to update.show() to show the confirmation dialogue
after reachin GOT_ZIP status.

Also, there was a segmentation fault when closing the application
because some classes don't have their full initialization in the
constructor and they need a refresh to fully initialize. This is
achieved by calling cothread_yield_irq().

Now that the test passes, the test can be run automatically in the
makefile.