Font: Fix rendering of characters on right edge of clip area
Sometimes the last character of the line was printed corrupted because it
doesn't fully fit in the clip area. The "data" pointer always increases by
one, even if some pixels are skipped. This patch recalculates the data
pointer per row so that all rows are drawn correctly even if some pixels
are skipped.