Font: TextArea: Allow printing partial lines of text
When printing a text area the doSingleChar() checks that characters fully
fit horizontally. If they don't, it tells the caller to advance one line.

However, in commit d33412eb5115 ("Font: TextArea: Prevent printing
characters outside of bounds") a new check was added so that any
character that doesn't fully fit vertically isn't printed.

This is a problem in two situations:

- When printing text areas that can't fit all text. It's better to display
  the last line cut than to not display it at all.
- When scrolling the screen canvas. It only tries to print the lines that
  have just appeared on the screen, so it regularly prints individual
  horizontal lines of text. With the changes of that commit many parts
  of the screen would remain white while scrolling.

This commit fixes the two issues.

Importantly, this still prevents buttons to print text outside of bounds
when it's too long.